requirements.txt
nltk==3.4.5
joblib==0.14.1
pandas==0.25.3
spacy==2.2.3
streamlit==0.53.0
psycopg2==2.8.3
SQLAlchemy==1.3.5
Pillow==7.0.0
numpy==1.18.1
scikit_learn==0.22.1
PyYAML==5.3
